+++ src/Server/InverterService/Constants.cs
@@ -1,0 +1,6 @@
+﻿namespace InverterMon.Server.InverterService;
+
+public static class Constants
+{
+    public const int StatusPollingFrequencyMillis = 1500;
+}