+++ src/changelog.md
@@ -1,2 +1,3 @@
 ## changelog
+- reduce status polling interval
 - remove pv power correction factor to avoid incorrect grid usage calculation